requirements.txt
ffmpeg_python==0.2.0
funasr==1.1.3
librosa==0.9.2
loguru==0.7.2
numpy==1.23.5
onnxruntime==1.18.0
onnxruntime_gpu==1.18.1
pydub==0.25.1
soundfile==0.12.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
